Move colorMap.js up the directory tree
