+++ renderer/components/colorMap.js
@@ -4,6 +4,7 @@ import {
 
 const colorMap = {
   anomaly: theme.colors.yellow9,
+  blocked: theme.colors.red7,
   reachable: theme.colors.green8,
   error: theme.colors.gray6
 }
